requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
lxml>=4.9.0
scikit-learn>=1.3.0
tqdm>=4.65.0
gensim>=4.3.0
requests>=2.28.0
vaderSentiment>=3.3.0
spacy>=3.5.0
flair>=0.13.0
transformers>=4.30.0
captum>=0.7.0
lime>=0.2.0
shap>=0.44.0
matplotlib>=3.7.0
seaborn>=0.13.0
scipy>=1.11.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
